Split i915 & xe, implement xe ioctls and memory usage
