+++ src/extract_gpuinfo_intel_xe.c
@@ -1,0 +1,96 @@
+/*
+ *
+ * Copyright (C) 2022 Maxime Schmitt <maxime.schmitt91@gmail.com>
+ *
+ * This file is part of Nvtop and adapted from igt-gpu-tools from Intel Corporation.
+ *
+ * Nvtop is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Nvtop is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with nvtop.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+#include "nvtop/device_discovery.h"
+#include "nvtop/extract_gpuinfo_common.h"
+#include "nvtop/extract_processinfo_fdinfo.h"
+#include "nvtop/time.h"
+
+#include "extract_gpuinfo_intel.h"
+
+#include <assert.h>
+#include <drm/drm.h>
+#include <drm/xe_drm.h>
+#include <errno.h>
+#include <fcntl.h>
+#include <stdio.h>
+#include <string.h>
+#include <sys/ioctl.h>
+#include <unistd.h>
+#include <uthash.h>
+
+// Copied from https://gitlab.freedesktop.org/mesa/mesa/-/blob/main/src/intel/common/intel_gem.h
+static inline int intel_ioctl(int fd, unsigned long request, void *arg) {
+  int ret;
+
+  do {
+    ret = ioctl(fd, request, arg);
+  } while (ret == -1 && (errno == EINTR || errno == EAGAIN));
+  return ret;
+}
+// End Copy
+
+// Copied from https://gitlab.freedesktop.org/mesa/mesa/-/blob/main/src/intel/common/xe/intel_device_query.c
+static void *xe_device_query_alloc_fetch(int fd, uint32_t query_id, uint32_t *len) {
+  struct drm_xe_device_query query = {
+      .query = query_id,
+  };
+  if (intel_ioctl(fd, DRM_IOCTL_XE_DEVICE_QUERY, &query))
+    return NULL;
+
+  void *data = calloc(1, query.size);
+  if (!data)
+    return NULL;
+
+  query.data = (uintptr_t)data;
+  if (intel_ioctl(fd, DRM_IOCTL_XE_DEVICE_QUERY, &query)) {
+    free(data);
+    return NULL;
+  }
+
+  if (len)
+    *len = query.size;
+  return data;
+}
+// End Copy
+
+void gpuinfo_intel_xe_refresh_dynamic_info(struct gpu_info *_gpu_info) {
+  struct gpu_info_intel *gpu_info = container_of(_gpu_info, struct gpu_info_intel, base);
+  struct gpuinfo_dynamic_info *dynamic_info = &gpu_info->base.dynamic_info;
+
+  if (gpu_info->card_fd) {
+    int32_t length = 0;
+    struct drm_xe_query_mem_regions *regions =
+        xe_device_query_alloc_fetch(gpu_info->card_fd, DRM_XE_DEVICE_QUERY_MEM_REGIONS, &length);
+    if (regions) {
+      for (unsigned i = 0; i < regions->num_mem_regions; i++) {
+        struct drm_xe_mem_region mr = regions->mem_regions[i];
+        if (mr.mem_class == DRM_XE_MEM_REGION_CLASS_VRAM) {
+          SET_GPUINFO_DYNAMIC(dynamic_info, total_memory, mr.total_size);
+          SET_GPUINFO_DYNAMIC(dynamic_info, used_memory, mr.used);
+          SET_GPUINFO_DYNAMIC(dynamic_info, free_memory, dynamic_info->total_memory - dynamic_info->used_memory);
+          break;
+        }
+      }
+      free(regions);
+    }
+  }
+}
